refactor: separe dto de update de eventos e de update de vagas de evento
