+++ apps/doc-flow/src/events/dto/update-event-vacancies.dto.ts
@@ -1,0 +1,13 @@
+import { ApiProperty } from '@nestjs/swagger';
+import { IsNumber } from 'class-validator';
+
+
+
+export class UpdateEventVacanciesDto {
+  @ApiProperty({
+    example: 10,
+    description: 'Vacancies',
+  })
+  @IsNumber()
+  vacancies: number;
+}
