+++ apps/doc-flow/src/events/dto/update-event.dto.ts
@@ -1,16 +1,4 @@
-import { PartialType } from '@nestjs/mapped-types';
-import { CreateEventDto } from './create-event.dto';
-import { ApiProperty } from '@nestjs/swagger';
-import { IsNumber } from 'class-validator';
+import { PartialType } from "@nestjs/swagger";
+import { CreateEventDto } from "./create-event.dto";
 
 export class UpdateEventDto extends PartialType(CreateEventDto) { }
-
-
-export class UpdateEventVacanciesDto {
-  @ApiProperty({
-    example: 10,
-    description: 'Vacancies',
-  })
-  @IsNumber()
-  vacancies: number;
-}
